No emergency sentence split in Unicode entities
Don't emergency sentence split if one of the last 10 chars is &, because
that means we might be in the middle of an encoded &U12345678; Unicode
entity, which should always be kept as part of one sentence.

NOTE: Entities are 11 chars, but if we've reached the final ;, we are
already free to split at that point. So we only need to check the last
10 chars for the presence of &.